Fix bug with ephemeral unmount
The juju config_get command will return an empty string if a
config option isn't set.  The ideal thing here is to change
the juju library to return Result<Option<String>, JujuError>
and set the Option to None in the case of empty strings. In
the mean time this is a workaround.