Format drives in parallel (#27)
* Format drives in parallel

This takes all the drives specified in juju storage or
manually and runs them through a series of checks to see
if they are a block device, if they are initialized, etc.
After all checks are done in sequential order they are
formatted in parallel.

* Only use initialized devices when setting up the volume

* Skip invalid block devices

* Make the tasks more parallel

* Use crossbeam scoped threads to check for rayon problems

* Use crossbeam instead of rayon.  Seems to actually be parallel

* Add logging to show which devices are being skipped

* Fix typo

* Try using futures instead

* Futures did not work.  Back to just spawning children

* Add/Remove entries from the fstab as needed

* Fix config_changed

After a deploy it became obvious that config_changed
is needed on the subordinate machines or they won't
format any of their drives.  This also adds fstab
entries for any drives it formats and mounts so
they can survive a reboot.